+++ src/lightning/fabric/utilities/imports.py
@@ -17,7 +17,10 @@ import operator
 import platform
 import sys
 
-from lightning_utilities.core.imports import compare_version
+from lightning_utilities.core.imports import RequirementCache, compare_version
+
+_NUMPY_AVAILABLE = RequirementCache("numpy")
+
 
 _IS_WINDOWS = platform.system() == "Windows"